refactor(rpc): collapse the #579 identity-resolver contract into one canonical definition
Move IdentityKeys, IdentityResolver, Ed25519Vk, MlDsaVk from auth::atproto_perimeter
(where they were a forward-declared 'Mirror' placeholder) into crate::identity as the
single canonical #579 contract. The perimeter gateway and mesh admission are generic
over the trait and unchanged; only the definition site moves. Re-exported from
auth::mod so existing crate::auth paths keep working. Removes the duplication that
would have drifted once #579 lands its real dispatcher.
